fix: session was not closed issue
Hi Guys,
I fixed error in Android!

In version 0.6.0, the app turned off in 30 times.
In version 0.6.1, the app turned off in 70 times.
and Now, The app does not turn off
The animation continued to be soft and the amount of heat generated became very low.
Overall, it was optimized.